Remove c-ares link step
c-ares is a transitive, public dependency of `grpc::grpc++`. As a result of this,
libraries linking against grpc also link against `c-ares`.

Utilizing this relationship between c-ares and grpc also simplifies the code in a sense
that we don't need to expose c-ares as a transitive dependency in the `carbon-grpc` export
target configuration.