Fix install rules
There are a number of factors at work here. Firstly, we were directly
installing a file generated from the `export` command.

Quoting the `export` command documentation`:
```
The file created by this command is specific to the build tree and
should never be installed
```

The second issue at play is that a `DESTINATION` parameter was specified
with the `install(TARGET ...)` command. This command does not have a
`DESTINATION` parameter, thus there's no value in passing a destination
with this installation rule.